# docs/swagger.yaml
basePath: /
definitions:
  github.com_miprokop_fication_internal_http-server_handlers.errorResponse:
    properties:
      message:
        type: string
    type: object
  github.com_miprokop_fication_internal_http-server_handlers.eventAllData:
    properties:
      created_by_id:
        type: string
      creation_date:
        type: string
      description:
        type: string
      end_date:
        type: string
      event_status:
        type: string
      event_type:
        type: string
      id:
        type: string
      image_path:
        type: string
      name:
        type: string
      organization_id:
        type: string
      staff:
        items:
          $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.staffEvents'
        type: array
      steps:
        items:
          $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.stepShortResponse'
        type: array
    type: object
  github.com_miprokop_fication_internal_http-server_handlers.eventRequest:
    properties:
      created_by_id:
        type: string
      creation_date:
        type: string
      description:
        type: string
      end_date:
        type: string
      event_status:
        type: string
      event_type:
        type: string
      id:
        type: string
      image_path:
        type: string
      name:
        type: string
      organization_id:
        type: string
      staff:
        items:
          $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.staffEvents'
        type: array
      steps:
        items:
          $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.stepRequest'
        type: array
    type: object
  github.com_miprokop_fication_internal_http-server_handlers.eventRequestUpdate:
    properties:
      created_by_id:
        type: string
      creation_date:
        type: string
      description:
        type: string
      end_date:
        type: string
      event_status:
        type: string
      event_type:
        type: string
      id:
        type: string
      image_path:
        type: string
      name:
        type: string
      organization_id:
        type: string
    type: object
...
  models.StaffID:
    properties:
      staff_id:
        type: string
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: API Server for challenge service
  title: Acheer API
  version: "1.0"
paths:
  /api/event/:
    post:
      consumes:
      - application/json
      description: |-
        create models.Event
        if no org id in request
        use a default org id
        if no type use public type
        event status by default process
        add a user who created this event to a member of this event
        if some steps there creates it
      operationId: create-event
      parameters:
      - description: event info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.eventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create event
      tags:
      - events
  /api/event/:id:
    delete:
      consumes:
      - application/json
      description: |-
        delete event by id
        also delete chained steps
      operationId: delete-event
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Event By ID
      tags:
      - events
    get:
      consumes:
      - application/json
      description: Get all invitations by current user
      operationId: get-invite-by-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.eventAllData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get event by id
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Update event by ID
      operationId: update-event
      parameters:
      - description: event
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.eventRequestUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.eventAllData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update event by id
      tags:
      - events
  /api/event/invitation/:
    get:
      consumes:
      - application/json
      description: Get all invitations by current user
      operationId: get-invites
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.staffEvents'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all invitations
      tags:
      - events
  /api/event/invitation/:id:
    post:
      consumes:
      - application/json
      description: |-
        Answer invite by ID
        status can be only models.InviteStatus
      operationId: answer-invite
      parameters:
      - description: staff and invites relation
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.staffEvents'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Answer invite
      tags:
      - events
  /api/event/invite/:id:
    post:
      consumes:
      - application/json
      description: Assign staff array
      operationId: assign-staff-to-event
      parameters:
      - description: staff ids and events
        in: body
        name: input
        required: true
        schema:
          items:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.staffEvents'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Assign Staff
      tags:
      - events
  /api/event/remove/:id:
    put:
      consumes:
      - application/json
      description: Remove staff from event and connected steps
      operationId: event-remove-staff
      parameters:
      - description: staffID
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.StaffID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Remove staff from event
      tags:
      - events
  /api/event/score/:id:
    get:
      consumes:
      - application/json
      description: get staff score in event by ID
      operationId: delete-event
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Staff Score In Event
      tags:
      - events
  /api/event/team/:id:
    get:
      consumes:
      - application/json
      description: Get all team's events by ID
      operationId: team-events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.eventAllData'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all events in team
      tags:
      - events
...
  /api/staff/:role:
    get:
      consumes:
      - application/json
      description: Get Events where staff takes a part by role
      operationId: get-staff-events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.eventsResponse'
        "400":
          description: Bad Request
          schema:
            type: ""
        "403":
          description: Forbidden
          schema:
            type: ""
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Events where staff takes a part by role
      tags:
      - events
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"